AI/requirements.txt
flask==2.3.2
numpy>=1.22,<1.25
pandas==2.0.3
tensorflow==2.13.0

requests~=2.31.0
cryptography~=41.0.4
python-dotenv~=1.0.0
boto3~=1.28.56
schedule~=1.2.0
nltk~=3.8.1
transformers~=4.33.2
joblib~=1.3.1